+++ src/pages/PostsManagerPage.tsx
@@ -1,30 +1,13 @@
 import { useEffect, useState } from "react"
 import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
 import { useLocation, useNavigate } from "react-router-dom"
-import {
-  Button,
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  Input,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-  Textarea,
-} from "../shared/ui"
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../shared/ui/table/table.tsx"
+import { Button } from "../shared/ui/button/button.tsx"
+import { Card, CardContent, CardHeader, CardTitle } from "../shared/ui/card/card.tsx"
+import { Input } from "../shared/ui/input/input.tsx"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../shared/ui/select/select.tsx"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../shared/ui/dialog/dialog.tsx"
+import { Textarea } from "../shared/ui/textarea/textarea.tsx"
 
 const PostsManager = () => {
   const navigate = useNavigate()
@@ -268,7 +251,6 @@ const PostsManager = () => {
   // 댓글 좋아요
   const likeComment = async (id, postId) => {
     try {
-
       const response = await fetch(`/api/comments/${id}`, {
         method: "PATCH",
         headers: { "Content-Type": "application/json" },
@@ -277,7 +259,14 @@ const PostsManager = () => {
       const data = await response.json()
       setComments((prev) => ({
         ...prev,
-        [postId]: prev[postId].map((comment) => (comment.id === data.id ? {...data, likes: comment.likes + 1} : comment)),
+        [postId]: prev[postId].map((comment) =>
+          comment.id === data.id
+            ? {
+                ...data,
+                likes: comment.likes + 1,
+              }
+            : comment,
+        ),
       }))
     } catch (error) {
       console.error("댓글 좋아요 오류:", error)
