+++ src/shared/ui/textarea/textarea.tsx
@@ -1,0 +1,12 @@
+import * as React from "react"
+
+// 텍스트 영역 컴포넌트
+export const Textarea = ({ className, ref, ...props }: React.ComponentPropsWithRef<"textarea">) => {
+  return (
+    <textarea
+      className={`flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+      ref={ref}
+      {...props}
+    />
+  )
+}
